+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.test.js
@@ -21,7 +21,10 @@ import { mockHTMLElement } from '../../global/js/utils/test-helper';
 const blockClass = `${pkg.prefix}--page-header`;
 
 /* Test properties. */
+const actionBarOverflowLabel = 'Show additional action bar items';
+
 const actionBarItems = [1, 2, 3, 4].map((item) => ({
+  key: `a-key-${item}`,
   renderIcon: Lightning16,
   iconDescription: `Action ${item}`,
   onClick: () => {},
@@ -41,6 +44,8 @@ const actionBarItemsNodes = (
 );
 
 const availableSpace = <span className="page-header-test--available-space" />;
+const breadcrumbOverflowLabel =
+  'Open and close additional breadcrumb item list.';
 const breadcrumbItems = (
   <>
     <BreadcrumbItem href="#">Breadcrumb 1</BreadcrumbItem>
@@ -51,16 +56,19 @@ const breadcrumbItems = (
 const classNames = ['client-class-1', 'client-class-2'];
 const pageActions = [
   {
+    key: '1',
     kind: 'secondary',
     label: 'Secondary button',
     onClick: () => {},
   },
   {
+    key: '2',
     kind: 'primary',
     label: 'Primary button',
     onClick: () => {},
   },
 ];
+const pageActionsOverflowLabel = 'Page actions...';
 
 const pageActionsDepTest = pageActions.map(({ label, ...rest }, index) => (
   <Button {...rest} key={index}>
@@ -80,7 +88,7 @@ const pageActionsDepTest2 = (
 
 const subtitle = 'Optional subtitle if necessary';
 const navigation = (
-  <Tabs data-testid="tabs">
+  <Tabs data-test-id="tabs">
     <Tab label="Tab 1" />
     <Tab label="Tab 2" />
     <Tab label="Tab 3" />
@@ -90,7 +98,7 @@ const navigation = (
 
 // supply enough tags to trigger TagSet overflow required props
 const tags = Array.from({ length: 20 }, () => ({
-  'data-testid': 'tags',
+  'data-test-id': 'tags',
   type: 'blue',
   label: 'A tag',
 }));
@@ -151,12 +159,15 @@ const testSizes = (el, property, _default) => {
 
 const testProps = {
   actionBarItems,
+  actionBarOverflowLabel,
   availableSpace,
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   className: classNames.join(' '),
   navigation,
   pageActions,
+  pageActionsOverflowLabel,
   subtitle,
   tags,
   title: titleObj,
@@ -191,7 +202,7 @@ describe('PageHeader', () => {
 
     mocks.push({
       id: 'uuidv4',
-      mock: uuidv4.mockImplementation(() => 'testid'),
+      mock: uuidv4.mockImplementation(() => 'test-id'),
     });
     window.scrollTo = jest.spyOn(window, 'scrollTo').mockImplementation();
     window.ResizeObserver = jest.fn().mockImplementation(() => ({
@@ -212,11 +223,11 @@ describe('PageHeader', () => {
   });
 
   test('renders an empty header when no props are set', () => {
-    const dataTestid = uuidv4();
-    render(<PageHeader data-testid={dataTestid} />);
+    const dataTestId = uuidv4();
+    render(<PageHeader data-test-id={dataTestId} />);
 
     // console.dir(screen.getByRole('region')); // section should be a region https://fae.disability.illinois.edu/rulesets/ROLE_5/
-    const header = screen.getByTestId(dataTestid);
+    const header = screen.getByTestId(dataTestId);
     expect(header).toHaveClass(blockClass);
 
     expect(header).toHaveClass(`${blockClass}--show-background`);
@@ -230,8 +241,6 @@ describe('PageHeader', () => {
     expect(
       document.querySelectorAll('span.page-header-test--available-space')
     ).toHaveLength(0);
-    expect(screen.queryAllByTestId('breadcrumbitem')).toHaveLength(0);
-    expect(screen.queryAllByTestId('tabs')).toHaveLength(0);
     expect(
       document.querySelectorAll(`.${blockClass}__page-actions`)
     ).toHaveLength(0);
@@ -239,7 +248,6 @@ describe('PageHeader', () => {
       0
     );
     expect(screen.queryByText(subtitle)).toBeNull();
-    expect(screen.queryAllByTestId('tags')).toHaveLength(0);
     expect(document.querySelectorAll(`.${blockClass}__title`)).toHaveLength(0);
     expect(screen.queryByText(titleObj.text)).toBeNull();
     expect(
@@ -248,11 +256,11 @@ describe('PageHeader', () => {
   });
 
   test('renders all the appropriate content when all props are set', () => {
-    const dataTestid = uuidv4();
-    render(<PageHeader {...testProps} data-testid={dataTestid} />);
+    const dataTestId = uuidv4();
+    render(<PageHeader {...testProps} data-test-id={dataTestId} />);
 
     // console.dir(screen.getByRole('region')); // section should be a region https://fae.disability.illinois.edu/rulesets/ROLE_5/
-    const header = screen.getByTestId(dataTestid);
+    const header = screen.getByTestId(dataTestId);
     expect(header).toHaveClass(blockClass);
 
     expect(header).toHaveClass(`${blockClass}--show-background`);
@@ -301,7 +309,12 @@ describe('PageHeader', () => {
   test('copes with actionBarItems as nodes', () => {
     const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
 
-    render(<PageHeader actionBarItems={actionBarItemsNodes} />);
+    render(
+      <PageHeader
+        actionBarItems={actionBarItemsNodes}
+        actionBarOverflowLabel={actionBarOverflowLabel}
+      />
+    );
 
     expect(warn).toBeCalledWith(
       'The usage of the prop `actionBarItems` of `PageHeader` has been changed and support for the old usage will soon be removed. Expects an array of objects with the following properties: iconDescription, renderIcon and onClick.'
@@ -313,7 +326,12 @@ describe('PageHeader', () => {
   test('with deprecated page actions', () => {
     const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
 
-    render(<PageHeader pageActions={pageActionsDepTest} />);
+    render(
+      <PageHeader
+        pageActions={pageActionsDepTest}
+        pageActionsOverflowLabel={pageActionsOverflowLabel}
+      />
+    );
 
     screen.getByText('Primary button', {
       selector: `.${blockClass}__page-actions .${pkg.prefix}--button-set-with-overflow__button-container:not(.${pkg.prefix}--button-set-with-overflow__button-container--hidden) .${carbon.prefix}--btn`,
@@ -329,7 +347,12 @@ describe('PageHeader', () => {
   test('with deprecated page actions in a fragment', () => {
     const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
 
-    render(<PageHeader pageActions={pageActionsDepTest2} />);
+    render(
+      <PageHeader
+        pageActions={pageActionsDepTest2}
+        pageActionsOverflowLabel={pageActionsOverflowLabel}
+      />
+    );
 
     screen.getByText('Primary button', {
       selector: `.${blockClass}__page-actions .${pkg.prefix}--button-set-with-overflow__button-container:not(.${pkg.prefix}--button-set-with-overflow__button-container--hidden) .${carbon.prefix}--btn`,
@@ -343,9 +366,9 @@ describe('PageHeader', () => {
   });
 
   it('adds additional properties to the containing node', () => {
-    const dataTestid = uuidv4();
-    render(<PageHeader data-testid={dataTestid} />);
-    screen.getByTestId(dataTestid);
+    const dataTestId = uuidv4();
+    render(<PageHeader data-test-id={dataTestId} />);
+    screen.getByTestId(dataTestId);
   });
 
   it('forwards a ref to an appropriate node', () => {
@@ -355,14 +378,14 @@ describe('PageHeader', () => {
   });
 
   test('collapse button works', () => {
-    const dataTestid = uuidv4();
+    const dataTestId = uuidv4();
     render(
       <PageHeader
         {...testProps}
         collapseHeaderLabel="Toggle collapse"
         expandHeaderLabel="Toggle expand"
         collapseHeaderToggleWanted={true}
-        data-testid={dataTestid}
+        data-test-id={dataTestId}
       />
     );
 
@@ -398,8 +421,9 @@ describe('PageHeader', () => {
   });
 
   test('Title row renders when PageActions but no Title', () => {
-    const { pageActions } = testProps;
-    render(<PageHeader {...{ pageActions }} />);
+    const { pageActions, pageActionsOverflowLabel = pageActionsOverflowLabel } =
+      testProps;
+    render(<PageHeader {...{ pageActions, pageActionsOverflowLabel }} />);
 
     expect(
       document.querySelectorAll(`.${blockClass}__page-actions`)
@@ -419,7 +443,11 @@ describe('PageHeader', () => {
 
   test('Title row renders when Title with pageActions and navigation but no subtitle or available space', () => {
     const { title } = testProps;
-    render(<PageHeader {...{ title, pageActions, navigation }} />);
+    render(
+      <PageHeader
+        {...{ title, pageActions, pageActionsOverflowLabel, navigation }}
+      />
+    );
 
     expect(
       document.querySelectorAll(
@@ -474,11 +502,13 @@ describe('PageHeader', () => {
 
   test('Without background', () => {
     const { title } = testProps;
+
     render(
       <PageHeader
         {...{
           title,
           background: false,
+          breadcrumbOverflowLabel: 'Show the breadcrumb overflow',
           breadcrumbItems,
         }}
         aria-label="Page header" // gives section role 'region'
@@ -486,5 +516,51 @@ describe('PageHeader', () => {
     );
 
     screen.getByRole('region');
+  });
+
+  test('ActionBar without overflow aria label', () => {
+    const error = jest.spyOn(console, 'error').mockImplementation(() => {});
+
+    const { title } = testProps;
+    render(
+      <PageHeader
+        {...{
+          title,
+          actionBarItems,
+        }}
+        aria-label="Page header" // gives section role 'region'
+      />
+    );
+
+    expect(error).toBeCalledWith(
+      expect.stringMatching(
+        /^Warning: Failed prop type: The prop `actionBarOverflowLabel` is marked as required in `PageHeader`/
+      )
+    );
+
+    jest.spyOn(console, 'error').mockRestore();
+  });
+
+  test('Breadcrumb without overflow aria label', () => {
+    const error = jest.spyOn(console, 'error').mockImplementation(() => {});
+
+    const { title } = testProps;
+    render(
+      <PageHeader
+        {...{
+          title,
+          breadcrumbItems,
+        }}
+        aria-label="Page header" // gives section role 'region'
+      />
+    );
+
+    expect(error).toBeCalledWith(
+      expect.stringMatching(
+        /^Warning: Failed prop type: The prop `overflowAriaLabel` is marked as required in `BreadcrumbWithOverflow`/
+      )
+    );
+
+    jest.spyOn(console, 'error').mockRestore();
   });
 });