+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.stories.js
@@ -49,16 +49,17 @@ export default {
     actions: {
       control: {
         type: 'select',
-        options: {
-          'One button': 0,
-          'One button (ghost)': 1,
-          'Two buttons': 2,
-          'Three buttons with ghost': 3,
-          'Three buttons': 4,
-          None: 5,
+        labels: {
+          0: 'One button',
+          1: 'One button (ghost)',
+          2: 'Two buttons',
+          3: 'Three buttons with ghost',
+          4: 'Three buttons',
+          5: 'None',
         },
         default: 0,
       },
+      options: [0, 1, 2, 3, 4],
     },
     slideIn: {
       table: {
@@ -74,7 +75,7 @@ const defaultStoryProps = {
   title:
     'Incident management for your application, testing a very long title to see how this behaves with a longer title',
   subtitle:
-    'This is some text that would talk about how you could investigate incidednt management within this side panel.',
+    'This is some text that would talk about how you could investigate incident management within this side panel.',
 };
 
 const headerData = [